refactor: extracting routes to own component
